Fixing some dependency bugs.
Mainly, the runtime was spawning structured concurrent tasks which reference def'd vars. The tasks would attempt to look up the value of the var but get stuck on a countdownlatch because the main thread was not finished manifesting the namespace. However, then the main thread would wait for the task to finish and both would get stuck forever. A change was made to the privilege that the thread loading the namespace the first time enjoys, extending it to any spawned structured concurrent subtasks that thread spawns as well.